-chg: ??????????? ???????? ?? ?????????? ???????? ?? ???????????? ???????????: scintilla 04.06.2011 r3689 scite 31.05.2011 r3513 ????????? ????????? [reverse.find] ?.?. ??? ??????????? ?????? ? ??????????? ??????. ?????????? ??? ?? ????????????? ?? ? VC71, ?? ? MinGW (???????? ???????? ? ??????? 1793)
SciTEBase.cxx
..\src\SciTEBase.cxx(906) : error C2248: 'GUI::ScintillaWindow::ScintillaWindow' : cannot access private member declared in class 'GUI::ScintillaWindow'
        c:\TEMP\scite-ru\src\scite\src\GUI.h(143) : see declaration of 'GUI::ScintillaWindow::ScintillaWindow'
        c:\TEMP\scite-ru\src\scite\src\GUI.h(141) : see declaration of 'GUI::ScintillaWindow'
..\src\SciTEBase.cxx(906) : warning C4239: nonstandard extension used : 'initializing' : conversion from 'GUI::ScintillaWindow' to 'GUI::ScintillaWindow &'
        A reference that is not to 'const' cannot be bound to a non-lvalue
..\src\SciTEBase.cxx(906) : error C2248: 'GUI::ScintillaWindow::ScintillaWindow' : cannot access private member declared in class 'GUI::ScintillaWindow'
        c:\TEMP\scite-ru\src\scite\src\GUI.h(143) : see declaration of 'GUI::ScintillaWindow::ScintillaWindow'
        c:\TEMP\scite-ru\src\scite\src\GUI.h(141) : see declaration of 'GUI::ScintillaWindow'
..\src\SciTEBase.cxx(4132) : error C2248: 'GUI::ScintillaWindow::ScintillaWindow' : cannot access private member declared in class 'GUI::ScintillaWindow'
        c:\TEMP\scite-ru\src\scite\src\GUI.h(143) : see declaration of 'GUI::ScintillaWindow::ScintillaWindow'
        c:\TEMP\scite-ru\src\scite\src\GUI.h(141) : see declaration of 'GUI::ScintillaWindow'

SciTEIO.cxx
..\src\SciTEIO.cxx(1254) : error C2511: 'void SciTEBase::GrepRecursive(SciTEBase::GrepFlags,FilePath,const char *,const GUI::gui_char *,unsigned int)' : overloaded member function not found in 'SciTEBase'
        c:\TEMP\scite-ru\src\scite\src\SciTEBase.h(319) : see declaration of 'SciTEBase'
..\src\SciTEIO.cxx(1364) : error C2660: 'SciTEBase::GrepRecursive' : function does not take 5 arguments

+++ src/scite/src/SciTEBuffers.cxx
@@ -1496,7 +1496,7 @@ int DecodeMessage(const char *cdoc, char *sourcePath, int format, int &column) {
 				startPath++;
 				const char *endPath = strchr(startPath, '\"');
 				if (endPath) {
-					int length = endPath - startPath;
+					ptrdiff_t length = endPath - startPath;
 					if (length > 0) {
 						strncpy(sourcePath, startPath, length);
 						sourcePath[length] = 0;
@@ -1535,7 +1535,7 @@ int DecodeMessage(const char *cdoc, char *sourcePath, int format, int &column) {
 				start++;
 			}
 			const char *endPath = strchr(start, '(');
-			int length = endPath - start;
+			ptrdiff_t length = endPath - start;
 			if ((length > 0) && (length < MAX_PATH)) {
 				strncpy(sourcePath, start, length);
 				sourcePath[length] = 0;
@@ -1572,7 +1572,7 @@ int DecodeMessage(const char *cdoc, char *sourcePath, int format, int &column) {
 						space2--;
 					}
 
-					int length = space2 - space;
+					ptrdiff_t length = space2 - space;
 
 					if (length > 0) {
 						strncpy(sourcePath, space, length);
@@ -1587,7 +1587,7 @@ int DecodeMessage(const char *cdoc, char *sourcePath, int format, int &column) {
 			// perl
 			const char *at = strstr(cdoc, " at ");
 			const char *line = strstr(cdoc, " line ");
-			int length = line - (at + 4);
+			ptrdiff_t length = line - (at + 4);
 			if (at && line && length > 0) {
 				strncpy(sourcePath, at + 4, length);
 				sourcePath[length] = 0;
@@ -1796,7 +1796,7 @@ int DecodeMessage(const char *cdoc, char *sourcePath, int format, int &column) {
 				 */
 			const char *startPath = strrchr(cdoc, '(') + 1;
 			const char *endPath = strchr(startPath, ':');
-			int length = endPath - startPath;
+			ptrdiff_t length = endPath - startPath;
 			if (length > 0) {
 				strncpy(sourcePath, startPath, length);
 				sourcePath[length] = 0;
@@ -1812,7 +1812,7 @@ int DecodeMessage(const char *cdoc, char *sourcePath, int format, int &column) {
 			const char *startPath = cdoc + 4;
 			const char *endPath = strchr(startPath, '\t');
 			if (endPath) {
-				int length = endPath - startPath;
+				ptrdiff_t length = endPath - startPath;
 				strncpy(sourcePath, startPath, length);
 				sourcePath[length] = 0;
 				return 0;
@@ -1828,7 +1828,7 @@ bool SciTEBase::GoMessage(int dir) { //!-change-[GoMessageImprovement]
 	Sci_CharacterRange crange;
 	crange.cpMin = wOutput.Call(SCI_GETSELECTIONSTART);
 	crange.cpMax = wOutput.Call(SCI_GETSELECTIONEND);
-	int selStart = crange.cpMin;
+	long selStart = crange.cpMin;
 	int curLine = wOutput.Call(SCI_LINEFROMPOSITION, selStart);
 	int maxLine = wOutput.Call(SCI_GETLINECOUNT);
 	int lookLine = curLine + dir;
@@ -1858,7 +1858,7 @@ bool SciTEBase::GoMessage(int dir) { //!-change-[GoMessageImprovement]
 			SString message = GetRange(wOutput, startPosLine, startPosLine + lineLength);
 			char source[MAX_PATH];
 			int column;
-			int sourceLine = DecodeMessage(message.c_str(), source, style, column);
+			long sourceLine = DecodeMessage(message.c_str(), source, style, column);
 			if (sourceLine >= 0) {
 //!-start-[FindResultListStyle]
 				if (props.GetInt("lexer.errorlist.findliststyle", 1)&& '.' == source[0] && pathSepChar == source[1]) {
